+++ android/src/main/java/com/swmansion/reanimated/layoutReanimation/NativeMethodsHolder.java
@@ -4,6 +4,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 public interface NativeMethodsHolder {
-    public void startAnimationForTag(int tag, String type, HashMap<String, Integer> values);
+    public void startAnimationForTag(int tag, String type, HashMap<String, Float> values);
     public void removeConfigForTag(int tag);
 }
